requirements.txt
Flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
numpy==1.26.3
Pillow==10.2.0
torch==2.1.2
torchvision==0.16.2
